feat: setting a list of user login that can see Administration
This feature permit to set a list of user login that will be able to see the Administration link.
If the list is empty every users will be able to see the link (like before without this feature).
Maybe a start with CAS auth (or LDAP ?) to filter access to some part without to request the etherpad settings.